autogen-agents/requirements.txt
pyautogen>=0.2.0
python-dotenv>=1.0.0
openai>=1.0.0
